Convert to long ,

+++ print_integer_bonus.c
@@ -1,7 +1,7 @@
 #include "ft_printf_bonus.h"
 #include "libft/libft.h"
 
-static int	handle_minus_for_int(int number, t_tokens *tokens)
+static int	handle_minus_for_int(long number, t_tokens *tokens)
 {
 	int	number_len;
 	int	len;
